Move code to glue.rs
